Adding tests.
